Add files via upload
PEGAGUS IMPLEMENTATION ON BERKELEY - STANFORD WEBPAGES:
Graphs or networks are everywhere, ranging from the Internet Web graph, social networks (Facebook, Twitter), biological networks, and many more. The large volume of available data, the low cost of storage and the stunning success of online social networks and web2.0 applications all lead to graphs of unprecedented size. There are mainly two goals while analyzing these graphs or networks. One is to find patterns in these graphs and detect anomalies such as outliers, etc. Second is to scale up analyses to grab billions of nodes and edges.  Graph Mining is an area of data mining to find patterns, rules, and anomalies of graphs. Typical graph mining algorithms silently assume that the graph fits in the memory of a typical workstation or at least on a single disk; the above graphs violate these assumptions, spanning multiple Giga-bytes, and heading to Tera- and Peta- bytes of data. Here comes PEGASUS, based on Hadoop is a graph mining p